rename: 프로젝트 이름 [핸즈업]으로 변경
프로젝트 이름이 핸즈업으로 결정됨
