+++ app/src/androidTest/java/com/tenday/handsup/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.tenday.dojeon
+package com.tenday.handsup
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.tenday.dojeon", appContext.packageName)
+        assertEquals("com.tenday.handsup", appContext.packageName)
     }
 }
