+++ app/src/main/AndroidManifest.xml
@@ -10,13 +10,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Dojeon"
+        android:theme="@style/Theme.HandsUp"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.Dojeon">
+            android:theme="@style/Theme.HandsUp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
