+++ app/src/main/java/com/tenday/handsup/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.tenday.dojeon
+package com.tenday.handsup
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,14 +11,14 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.tenday.dojeon.ui.theme.DojeonTheme
+import com.tenday.handsup.ui.theme.HandsUpTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            DojeonTheme {
+            HandsUpTheme {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
                     Greeting(
                         name = "Android",
@@ -41,7 +41,7 @@ fun Greeting(name: String, modifier: Modifier = Modifier) {
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
-    DojeonTheme {
+    HandsUpTheme {
         Greeting("Android")
     }
 }
