+++ app/src/main/java/com/tenday/handsup/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.tenday.dojeon.ui.theme
+package com.tenday.handsup.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
@@ -34,7 +33,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun DojeonTheme(
+fun HandsUpTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
