+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Dojeon</string>
+    <string name="app_name">HANDSUP</string>
 </resources>
